pyproject.toml
[project]
name = "langextract-demo"
version = "0.1.0"
description = "Gemini-powered information extraction demo using LangExtract"
authors = [{ name = "Your Name", email = "your@email.com" }]
dependencies = [
    "langextract>=0.1.3",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.8"

[project.scripts]
langextract-demo = "langextract_demo.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
